+++ Analysis/TimeSeries/MultiSC.py
@@ -1,15 +1,8 @@
 #!/usr/bin/env python
-import sys, os
-sys.path.insert(0,os.environ['HOME']+'/AJGAR/TurbPlasma')
 import random as rnd
 import numpy as np
 import numpy.fft as nf
-import AnalysisFunctions as af
 import pandas as pd
-#import seaborn as sns
-import matplotlib.pyplot as plt
-from . import ftsa
-import numba
 
 pi = np.pi
