+++ Analysis/TimeSeries/OLLibs/F90/__init__.py
@@ -1,1 +1,2 @@
 from . import DataCleaning
+from .ftsa import *
